Updated ww130_test to work with WW130 code v0.5.3
As of this date the WW130 software is at v503 and the inter-processor communication algorithm has changed.
To make the ww130_test app work with WW130 V502 etc I have had to make a small change in ``com_task.c`` which is invoked when ``#define WW130_500`` is uncommented. In this case the PA0 pin is toggled by the Seeed board to attract the attention of the WW130.